refactor: 경로 분기, typo, 미사용 Import 제거 (#126)
