+++ src/main/java/com/api/ttoklip/domain/notification/aop/NotificationAspect.java
@@ -1,7 +1,7 @@
 package com.api.ttoklip.domain.notification.aop;
 
 import com.api.ttoklip.domain.notification.aop.annotation.SendNotification;
-import com.api.ttoklip.domain.notification.entity.PostETCEvent;
+import com.api.ttoklip.domain.notification.event.PostETCEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.JoinPoint;
