+++ src/main/java/com/api/ttoklip/domain/notification/event/PostETCEvent.java
@@ -1,0 +1,13 @@
+package com.api.ttoklip.domain.notification.event;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class PostETCEvent {
+
+    private final Long postId;
+    private final String className;
+    private final String methodName;
+}
